feat(explorer): configure apps in-panel inside the launcher (view swap)
Clicking Configure… no longer closes the launcher and opens a separate
full-screen dialog — the launcher body swaps in place to the config view
(back button + "Configure apps" header, Available/Selected panels,
reorder, Reset/Save/Cancel), inside the same command-surface panel.
Escape layers: config view -> clear filter -> close.

Mechanics:
- Extracted all config logic/content into a chrome-free
  mj-user-app-config-content component (explorer-settings): loads on
  init, batch-saves, signals hosts via Saved/Cancelled. Reorder helpers
  and persistence untouched.
- mj-user-app-config is now a thin full-screen-modal wrapper around the
  content with its public API (ShowDialog/Open/Close/ConfigSaved)
  unchanged — the Home dashboard consumer needs zero edits (regression
  verified live: opens, loads 26 rows, closes).
- Launcher (explorer-core): ConfigMode state + head/body/footer swap;
  post-save the grid refreshes reactively via the existing
  UserInfoEngine -> ApplicationManager event chain.

Passing fixes: dialog title icon fa-grid-2 (FA Pro, rendered blank) ->
fa-grip; modal backdrop rgba(0,0,0,0.5) -> var(--mj-bg-overlay).
Known flag: the config content CSS still uses legacy --mat-sys-*
variables (pre-existing; token-migration candidate).

Verified: explorer-settings 92/92, explorer-core 90/90 (4 new DOM tests:
view swap, back, Escape layering, save-returns-to-grid), color gates
clean, live-verified swap + Home dialog regression.
