+++ packages/Angular/Explorer/explorer-core/src/lib/shell/components/header/app-switcher.component.html
@@ -28,6 +28,23 @@
       aria-label="Application launcher"
       (keydown)="onPanelKeydown($event)">
 
+      @if (ConfigMode) {
+        <!-- Configuration view: same panel, body swapped -->
+        <div class="launcher-head launcher-head--config">
+          <button class="launcher-back" type="button" aria-label="Back to apps" (click)="ExitConfigMode()">
+            <i class="fa-solid fa-arrow-left" aria-hidden="true"></i>
+          </button>
+          <h2 class="launcher-title" id="mj-launcher-config-title">Configure apps</h2>
+          <button class="launcher-close" type="button" aria-label="Close application launcher" (click)="CloseLauncher()">
+            <i class="fa-solid fa-xmark" aria-hidden="true"></i>
+          </button>
+        </div>
+        <mj-user-app-config-content
+          class="launcher-config-body"
+          (Saved)="ExitConfigMode()"
+          (Cancelled)="ExitConfigMode()">
+        </mj-user-app-config-content>
+      } @else {
       <div class="launcher-head">
         <div class="launcher-filter">
           <i class="fa-solid fa-magnifying-glass" aria-hidden="true"></i>
@@ -81,13 +98,14 @@
           <span><kbd>↵</kbd> open</span>
           <span><kbd>esc</kbd> close</span>
         </span>
-        <button class="launcher-configure" type="button" testId="app-switcher-configure" (click)="openConfigDialog()">
+        <button class="launcher-configure" type="button" testId="app-switcher-configure" (click)="EnterConfigMode()">
           <i class="fa-solid fa-gear" aria-hidden="true"></i>
           <span>Configure…</span>
         </button>
       </div>
 
       <span class="sr-only" role="status">{{ VisibleApps.length }} applications available</span>
+      }
     </div>
   }
 </div>
@@ -126,9 +144,3 @@
   </div>
 </ng-template>
 
-<!-- App Configuration Dialog -->
-<mj-user-app-config
-  #appConfigDialog
-  [(ShowDialog)]="showConfigDialog"
-  (ConfigSaved)="onConfigSaved()">
-</mj-user-app-config>
