+++ packages/Angular/Explorer/explorer-settings/src/lib/module.ts
@@ -35,6 +35,7 @@ import { SettingsComponent } from './settings/settings.component';
 // User-facing components
 import { UserProfileSettingsComponent } from './user-profile-settings/user-profile-settings.component';
 import { UserAppConfigComponent } from './user-app-config/user-app-config.component';
+import { UserAppConfigContentComponent } from './user-app-config/user-app-config-content.component';
 import { NotificationPreferencesComponent } from './notification-preferences/notification-preferences.component';
 
 // New user settings components
@@ -63,6 +64,7 @@ import { ApplicationDialogComponent } from './application-management/application
     // User-facing components
     UserProfileSettingsComponent,
     UserAppConfigComponent,
+    UserAppConfigContentComponent,
     NotificationPreferencesComponent,
     // New user settings components
     GeneralSettingsComponent,
@@ -113,6 +115,7 @@ import { ApplicationDialogComponent } from './application-management/application
     // User-facing components
     UserProfileSettingsComponent,
     UserAppConfigComponent,
+    UserAppConfigContentComponent,
     NotificationPreferencesComponent,
     // New user settings components
     GeneralSettingsComponent,
